added Register feature, config and controller structs. Cleaned up controllers and models. Added data folder to take care of database queries and actions.

+++ src/models/models.go
@@ -6,7 +6,7 @@ import (
 
 type (
 	User struct {
-		User_ID      int           `json:"user_id"`
+		User_ID      int64           `json:"user_id"`
 		Username     string        `json:"username"`
 		Email        string        `json:"email"`
 		Password     string        `json:"password"`
